harmonize closure property paths
In #37, we introduced the possibility to define the read property path
as a closure. This was already possible for write operations too. The
difference was that closures defined for the write property path did
not call any methods of the model themselves, but only returned another
property path.

This is changed now and the configured closures need to call model
methods just like when reading data from the model. Wrongly submitted
data must now be indicated by throwing exceptions from inside the
anonymous functions (or by forwarding them from the model).