forbid the data_class option on immutable forms
The view data of an immutable form is an array. Using the "data_class"
option means that a new object will be instantiated breaking the data
mapping process. But since the Form component never needs to create new
objects when the initial data is empty (this case is handled by the
ValueObjectTransformer) configuring it is useless anyway.